wip(remarq): pandocify manuscript files
